requirements.txt
requests~=2.28.1
pandas~=1.5.1
python-dotenv~=0.21.0